pyproject.toml
[virtualenvs]
create = false

[tool.poetry]
name = "crypto-sentiment-analyzer"
version = "0.1.0"
description = "A cryptocurrency sentiment analysis tool using Reddit, RSS feeds, and Fear & Greed Index"
authors = ["Glenn Steven Santoso <glennstevensantoso@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = ">=2.31.0"
python-dotenv = ">=1.0.0"
nltk = ">=3.8.1"
pandas = ">=2.1.0"
numpy = ">=1.24.3"
matplotlib = ">=3.7.1"
seaborn = ">=0.12.2"
google-api-python-client = ">=2.0.0"
google-auth-httplib2 = ">=0.1.0"
google-auth-oauthlib = ">=0.4.0"
pytz = ">=2024.1"
praw = ">=7.7.1"
websocket-client = ">=1.6.0"
feedparser = "^6.0.10"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
analyze = "sentiment.reddit_analyzer:main"